packages/sutra-control/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.59.0
grpcio-tools==1.59.0
pydantic==2.5.0
numpy==1.24.3
protobuf==4.25.0